Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/flow-content-0/dialog-expected.txt




FAIL Closed dialog in width: 540px iframe assert_equals: right expected "0px" but got "auto"
FAIL Open dialog in width: 540px iframe assert_equals: top expected "8px" but got "auto"
FAIL Modal dialog in width: 540px iframe assert_equals: top expected "56px" but got "auto"
FAIL Closed dialog in width: 538px iframe assert_equals: right expected "0px" but got "auto"
FAIL Open dialog in width: 538px iframe assert_equals: top expected "8px" but got "auto"
FAIL Modal dialog in width: 538px iframe assert_equals: top expected "56px" but got "auto"
